+++ NEWS.md
@@ -59,6 +59,11 @@ reduce2_right(.x = letters[1:4], .y = paste2, .f = c("-", ".", "-")) # working
 
 ## Minor improvements and fixes
 
+* Fixed ordering of names returned by `accumulate_right()`
+  output. They now correspond to the order of inputs.
+
+* Fixed names of `accumulate()` output when `.init` is supplied.
+
 * New `modify2()` and `imodify()` functions. These work like `map()`
   and `imap()` but preserve the type of `.x` in the return value.